Create WinSCP_SFTP_Backup.ps1
Copy file from remote Windows server to local virtual C drive (Ideally this should be changed to a network drive specifically for backups). Change file name and upload file SFTP to linux server using encrypted credentials form XML file. If successful send a text to me and email Jira. This will create a ticket within Jira based of the template provided. Sender will be Audit officer and the file will be included as an attachment. If the script fails, send me a text and email with the error code.